+++ test/latin1/Main.hs
@@ -1,30 +1,28 @@
 {-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE TypeFamilies   #-}
 
 module Main (main) where
 
 import           Control.Exception.Safe (bracket)
-import           Control.Monad (join)
-import           Data.Char (GeneralCategory (NotAssigned, PrivateUse),
-                            generalCategory)
+import qualified Data.ByteString as BS
 import           Data.Foldable (traverse_)
 import           Data.Functor.Identity (Identity)
 import           Data.Int (Int64)
 import           Data.Kind (Type)
-import           Data.Text (Text, chunksOf, filter, pack)
-import           Data.Vector (Vector, fromList, imap, toList)
+import           Data.Text (Text)
+import           Data.Text.Encoding (decodeLatin1)
+import           Data.Vector (Vector)
+import qualified Data.Vector as V
 import           Database.Beam (Beamable, Columnar, Database, DatabaseSettings,
                                 Table (PrimaryKey, primaryKey), TableEntity,
-                                defaultDbSettings)
-import           Database.Beam.MySQL (MySQL, ViaJson (ViaJson), runBeamMySQL)
-import           Database.Beam.Query (QExpr, SqlInsertValues, all_, asc_,
-                                      insert, insertValues, orderBy_, runInsert,
-                                      runSelectReturningList, select)
+                                all_, asc_, defaultDbSettings, insert,
+                                insertValues, orderBy_, runInsert,
+                                runSelectReturningList, select)
+import           Database.Beam.MySQL (MySQL, runBeamMySQL)
+import           Database.Beam.Query (QExpr, SqlInsertValues)
 import           Database.MySQL.Base (MySQLConn, Query (Query), close, connect,
                                       execute_)
 import           Database.MySQL.Temp (MySQLDB, toConnectInfo, withTempDB)
 import           GHC.Generics (Generic)
-import           Prelude hiding (filter)
 import           Test.Hspec (Spec, describe, hspec, it, shouldBe)
 
 main :: IO ()
@@ -32,70 +30,55 @@ main = do
   res <- withTempDB go
   hspec . spec $ res
   where
-    go :: MySQLDB -> IO (Vector (Text, Text))
+    go :: MySQLDB -> IO (Vector Text)
     go db = bracket (connect . toConnectInfo $ db)
                     close
                     (\conn -> setUpDB conn >> insertSample conn >> retrieveSample conn)
 
 -- Helpers
 
-spec :: Vector (Text, Text) -> Spec
-spec v = describe "Unicode round-tripping" $ do
+spec :: Vector Text -> Spec
+spec v = describe "Latin-1-only valid round-tripping" $ do
   it "should round-trip properly" $ do
     v `shouldBe` sampleData
 
--- every valid Unicode code point, in chunks of 16
-sampleData :: Vector (Text, Text)
-sampleData =
-  fmap (join (,)) .
-  fromList .
-  chunksOf 16 .
-  filter go .
-  pack $ [minBound .. maxBound]
-  where
-    go :: Char -> Bool
-    go c = case generalCategory c of
-      NotAssigned -> False
-      PrivateUse  -> False
-      _           -> True
+-- Every byte from 0x80 to 0xFF as a singleton string
+sampleData :: Vector Text
+sampleData = V.fromList . fmap (decodeLatin1 . BS.singleton) $ [0x80 .. 0xFF]
 
 setUpDB :: MySQLConn -> IO ()
 setUpDB conn = traverse_ (execute_ conn) [
   "create database test;",
   "use test",
-  makeTestTable]
+  makeTestTable
+  ]
   where
     makeTestTable :: Query
     makeTestTable = Query $
       "create table test_table (" <>
       "id bigint primary key not null, " <>
-      "data varchar(255) not null, " <>
-      "json varchar(255) not null);"
+      "data varchar(255) not null);"
 
 insertSample :: MySQLConn -> IO ()
 insertSample conn = runBeamMySQL conn . runInsert . insert (_testTestTable testDB) $ go
   where
     go :: SqlInsertValues MySQL (TestT (QExpr MySQL s))
-    go = insertValues . toList . imap buildTestT $ sampleData
-    buildTestT :: Int -> (Text, Text) -> TestT Identity
-    buildTestT i (t, t') = TestT (fromIntegral i) t (ViaJson t')
+    go = insertValues . V.toList . V.imap buildTestT $ sampleData
+    buildTestT :: Int -> Text -> TestT Identity
+    buildTestT i t = TestT (fromIntegral i) t
 
-retrieveSample :: MySQLConn -> IO (Vector (Text, Text))
+retrieveSample :: MySQLConn -> IO (Vector Text)
 retrieveSample conn =
-  fmap (fmap go . fromList) . runBeamMySQL conn . runSelectReturningList . select $ do
+  fmap V.fromList . runBeamMySQL conn . runSelectReturningList . select $ do
     res <- orderBy_ (asc_ . _testId) . all_ . _testTestTable $ testDB
-    pure (_testData res, _testJson res)
-  where
-    go :: (Text, ViaJson Text) -> (Text, Text)
-    go (t, ViaJson t') = (t, t')
+    pure . _testData $ res
 
 -- Beam boilerplate
 
 data TestT (f :: Type -> Type) = TestT
   {
     _testId   :: Columnar f Int64,
-    _testData :: Columnar f Text,
-    _testJson :: Columnar f (ViaJson Text)
+    _testData :: Columnar f Text
   }
   deriving stock (Generic)
   deriving anyclass (Beamable)